+++ CHANGELOG.md
@@ -27,6 +27,9 @@
 * Fix: [#430](https://github.com/savonrb/savon/pull/430) allows you to rescue and ignore
   `Savon::Error` errors in production while still having mocks trigger test failures.
 
+* Fix: [#393](https://github.com/savonrb/savon/pull/393) changed `Savon::SOAPFault` to work
+  with generic response Hash keys.
+
 ### 2.1.0 (2013-02-03)
 
 * Feature: [#372](https://github.com/savonrb/savon/pull/372) added global `ssl_cert_key_password` option.